requirements.txt
pandas==2.1.0
openai==1.0.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
